+++ tests/integration/legacy/test_base_sw_legacy.py
@@ -4,7 +4,7 @@
 #
 # Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 
-from ..test_base_sw_headers_attrs import TestBaseSw
+from ..test_base_sw import TestBaseSw
 
 
 class TestBaseSwLegacy(TestBaseSw):
